Suppress non-null property initialization warnings on DbSet properties, by acknowledging that we *know* EF will initialize them for us.
The warning was "Non-nullable property 'Contact' is uninitialized. Consider declaring the property as nullable."

However, we know that EF Core guarantees non-null values for the DbSet properties of a DbContext, so instead we initialize to default (technically null) in combination with the null-forgiveness operator (!). This way, the rest of the system that *uses* the property can safely assume that it will never be null.

This approach follows guidance from the documentation: https://docs.microsoft.com/en-us/ef/core/miscellaneous/nullable-reference-types with a superficial replacement of "default!" for "null!" for readability.